+++ mods/BrewLAN_Units/BrewResearch/hook/lua/aibrain.lua
@@ -13,18 +13,30 @@ AIBrain = Class(AIBrain) {
                 local fname = {'UEF', 'Aeon', 'Cybran', 'Seraphim'}--[f]
                 fname = fname[f]
 
-                --Generate the full research list
-                self.BrewRND.ResearchList = {}
+                --Generate the full research lists
+                self.BrewRND.ResearchList = {}      --Populated with all research options for the faction now
+                                                    --Looks like an array of id's
+                                                    --Items are removed as they are finished.
+
+                self.BrewRND.CategoryResearch = {}  --Populated with all category research options now
+                                                    --keyed with id's values are categories
+                                                    --Items are removed as they are requested, or finished if they weren't requested.
+
                 for id, bp in __blueprints do
                     if bp.Categories then
                         if bp.General.FactionName == fname and table.find(bp.Categories, 'BUILTBYRESEARCH') and not table.find(self.BrewRND.ResearchList, bp.BlueprintId) then
                             table.insert(self.BrewRND.ResearchList, bp.BlueprintId)
+                            if not __blueprints[bp.ResearchId] then
+                                self.BrewRND.CategoryResearch[bp.BlueprintId] = bp.ResearchId
+                            end
                         end
                     end
                 end
 
                 --Initialise the table to recieve requests.
-                self.BrewRND.ResearchRequests = {}
+                self.BrewRND.ResearchRequests = {}  --Populated with research items as other managers need things.
+                                                    --Looks like an array of id's
+                                                    --Items are removed as they are finished.
             end
         end,
 
@@ -34,6 +46,8 @@ AIBrain = Class(AIBrain) {
             local AIUtils = import('/lua/ai/aiutilities.lua')
             local pos = {ScenarioInfo.size[1]/2, 0, ScenarioInfo.size[2]/2}
             local radius = ScenarioInfo.size[1]
+
+            --This section could be updated to use the self.BrewRND.CategoryResearch list
 
             -- variable for tracking max tech level owned
             local maxtech = 0
@@ -111,16 +125,31 @@ AIBrain = Class(AIBrain) {
         --Request something be researched. Must be valid for this faction, and not already researched.
         AddResearchRequest = function(self, item)
             --if this request isn't already on the table, then...
-            if not table.find(self.BrewRND.ResearchRequests, item) and not table.find(self.BrewRND.ResearchRequests, item .. 'rnd') then
-                --if it directly is a valid research item add it
-                if table.find(self.BrewRND.ResearchList, item) then
+            local AddResearch = function(self, item)
+                if not table.find(self.BrewRND.ResearchRequests, item) and table.find(self.BrewRND.ResearchList, item) then
+                    WARN("AI research requesting " .. item)
                     table.insert(self.BrewRND.ResearchRequests, item)
-                -- if the unit code + rnd is a valid research item then add that
-                elseif table.find(self.BrewRND.ResearchList, item .. 'rnd') then
-                    table.insert(self.BrewRND.ResearchRequests, item .. 'rnd')
-                -- else we have un defined behaviour
-                else
-                    WARN("Unknown research request type " .. item)
+                    return true
+                end
+            end
+
+            -- Try it as written
+            if not AddResearch(self, item) then
+                -- Try it as with rnd
+                if not AddResearch(self, item .. 'rnd') then
+                    -- Try searching for a cat research
+                    local bp = __blueprints[item]
+                    if bp and bp.Categories then
+                        for id, cat in self.BrewRND.CategoryResearch do
+                            if table.find(bp.Categories, cat) then
+                                if AddResearch(self, id) then
+                                    --if this this works, nil the entry so we don't search for this cat research again.
+                                    --The request wont go away until it's completed.
+                                    self.BrewRND.CategoryResearch[id] = nil
+                                end
+                            end
+                        end
+                    end
                 end
             end
         end,
@@ -133,6 +162,7 @@ AIBrain = Class(AIBrain) {
                     table.remove(array, f)
                 end
             end
+            self.BrewRND.CategoryResearch[item] = nil
         end,
     }
 }